LayoutTests/imported/w3c/web-platform-tests/feature-policy/reporting/usb-reporting.https-expected.txt

CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating 'navigator.usb.getDevices')

Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'navigator.usb.getDevices')

NOTRUN USB Report Format
